見通しが悪くなってきたのでuser_handler.goにユーザー系を移す
