+++ web/server.go
@@ -6,7 +6,6 @@ import (
 	"html/template"
 	"io/ioutil"
 	"net/http"
-	"time"
 
 	"github.com/dimfeld/httptreemux"
 	"github.com/justinas/nosurf"
@@ -68,7 +67,7 @@ func (s *server) Handler() http.Handler {
 
 	handle("GET", "/", s.indexHandler())
 
-	// ユーザー登録・ログイン
+	// ユーザー登録・ログイン(user_handler.go)
 	handle("GET", "/signup", s.willSignupHandler())
 	handle("POST", "/signup", s.signupHandler())
 	handle("GET", "/signin", s.willSigninHandler())
@@ -92,85 +91,6 @@ func (s *server) indexHandler() http.Handler {
 		s.renderTemplate(w, r, "index.tmpl", map[string]interface{}{
 			"User": user,
 		})
-	})
-}
-
-// 登録ページ
-func (s *server) willSignupHandler() http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		s.renderTemplate(w, r, "signup.tmpl", nil)
-	})
-}
-
-// 登録処理
-func (s *server) signupHandler() http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		name, password := r.FormValue("name"), r.FormValue("password")
-		if err := s.app.CreateNewUser(name, password); err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		user, err := s.app.FindUserByName(name)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		expiresAt := time.Now().Add(24 * time.Hour)
-		token, err := s.app.CreateNewToken(user.ID, expiresAt)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		http.SetCookie(w, &http.Cookie{
-			Name:    sessionKey,
-			Value:   token,
-			Expires: expiresAt,
-		})
-		http.Redirect(w, r, "/", http.StatusSeeOther)
-	})
-}
-
-func (s *server) willSigninHandler() http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		s.renderTemplate(w, r, "signin.tmpl", nil)
-	})
-}
-
-func (s *server) signinHandler() http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		name, password := r.FormValue("name"), r.FormValue("password")
-		if ok, err := s.app.LoginUser(name, password); err != nil || !ok {
-			http.Error(w, "user not found or invalid password", http.StatusBadRequest)
-			return
-		}
-		user, err := s.app.FindUserByName(name)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		expiresAt := time.Now().Add(24 * time.Hour)
-		token, err := s.app.CreateNewToken(user.ID, expiresAt)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		http.SetCookie(w, &http.Cookie{
-			Name:    sessionKey,
-			Value:   token,
-			Expires: expiresAt,
-		})
-		http.Redirect(w, r, "/", http.StatusSeeOther)
-	})
-}
-
-func (s *server) signoutHandler() http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		http.SetCookie(w, &http.Cookie{
-			Name:    sessionKey,
-			Value:   "",
-			Expires: time.Unix(0, 0),
-		})
-		http.Redirect(w, r, "/", http.StatusSeeOther)
 	})
 }
